apps/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
httpx
anthropic
pandas
pyyaml
python-dotenv
pydantic
pydantic-settings
